app/requirements.txt
Flask==1.1.1
mysql-connector==2.2.9
scrapy==2.3